+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 2018-MM-DD v0.X.Y
 * Rename `ToJson/FromJson/ToCsv/FromCsv` -> `ToJSON/FromJSON/ToCSV/FromCSV`
+* qframe.Expr1 and qframe.Expr2 have been merged to one function qframe.Expr
 
 ### 2018-05-23 v0.1.0
 * Initial release